+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Added Recorder.RecordAsThis(), which allows objects to turn over their entire no
 ### Improved
 5-10% performance improvement for Recorder serialization.
 
+### Breaking
+Removed user-facing XML exposure, including Converter.ToXml, Converter.FromXml, and Recorder.Xml. This functionality wasn't being used by anyone I could find and was seriously complicating development. There is currently no replacement planned.
+
 ### Testing
 Add small tests.