+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 
-## [Unreleased]
+## [v0.3.0]
 ### Added
 Added Recorder.RecordAsThis(), which allows objects to turn over their entire node to a child object.